Documents/insurance-analytics/requirements.txt
et_xmlfile==2.0.0
numpy==2.4.3
openpyxl==3.1.5
pandas==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.17.0
tzdata==2025.3
